Change createFillings return type to function
This allows users to only call createFillings function and use it's
output directly within their code
